requirements.txt
apipkg==1.5
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
coverage==5.2.1
Django==3.1
django-crispy-forms==1.9.2
django-stubs==1.5.0
djangorestframework==3.11.1
djangorestframework-stubs==1.2.0
execnet==1.7.1
flake8==3.8.3
iniconfig==1.0.1
mccabe==0.6.1
more-itertools==8.4.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-forked==1.3.0
pytest-xdist==1.34.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
